Unify config-type triplet into shared/config-types
Move RawRepo, RawSubmoduleRepo, and isSectionMarker into a process-agnostic
src/shared/config-types.ts, re-exported from config-schema.ts and config-walk.ts
so existing importers keep working. Re-point the renderer Settings-modal files at
@shared/config-types so they no longer import runtime code from src/main, keeping
zod out of the SPA bundle.

Delete the stale config-walk.ts type subset and worktree/shared.ts's RawRepoExtended
interface; the canonical RawRepo now carries pinned_branch/install/env, so the unsafe
'as unknown as' double-cast is no longer needed.
